#6094 Add improvements found by review
